pyproject.toml
[tool.poetry]
name = "openai-discord-chatbot"
version = "0.1.0"
description = "A Discord bot that interacts using the OpenAI API, maintains conversation history, and utilizes system messages for context."
authors = ["Michael Taylor <90michaeltaylor@gmail.com>"]
readme = "README.md"
packages = [{include = "openai_discord_chatbot"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
openai = "^0.27.8"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
pydantic = {extras = ["dotenv"], version = "^2.2.1"}
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
